refactor(domains): restore tab navigation and shrink action buttons
Reverts the single-page layout back to separate tabs (Register, Transfer,
My Domains) for clearer navigation. Moves link-domain callout into the
My Domains tab and removes GSAP collapsible section. Shrinks search and
transfer buttons from full-width to content-sized.
